Fix links
Adapt links to RST structure. This is done through relative links
which works without having to change the content of the docs. Once
the automatic import is completed this can be turned into using
native references and labels in RST.

The glossary links use the native :term: role already.

There is an intermediate page serving as target for links which
can't be reflected in the current structure of the docs. These links
will have to be manually adjusted once the automatic import is done.
Then the page can be removed again.

Fixes #16.